Add compile and test functionality to CI server. Closes #6.
The CI server will now listen for webhooks. Requests are handled by cloning the branch in which the commit was made, and then building the maven project in that branch, which compiles and tests the project.

Server is now part of Maven project, and is run with "mvn exec:java"

Adds neccessary dependencies and plugins.

Adds instructions for setup, running and testing the CI server in README.

Co-authored-by: FelixSjogren <felixarvidsjogren@gmail.com>